+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/exporter.py
@@ -67,7 +67,7 @@ class Worker(threading.Thread):
 
     def run(self):  # pragma: NO COVER
         # Indicate that this thread is an exporter thread.
-        # Used to suppress tracking of requests in this thread
+        # Used to suppress tracking of requests in this thread.
         execution_context.set_is_exporter(True)
         src = self.src
         dst = self.dst