+++ contrib/opencensus-ext-flask/CHANGELOG.md
@@ -13,6 +13,8 @@ Released 2019-08-26
 
 - Updated `http.status_code` attribute to be an int.
   ([#755](https://github.com/census-instrumentation/opencensus-python/pull/755))
+- Fixes value for `http.route` in Flask middleware
+  ([#759](https://github.com/census-instrumentation/opencensus-python/pull/759))
 
 ## 0.7.1
 Released 2019-08-05